# docs/swagger.yaml
definitions:
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.LoginParams:
    properties:
      email:
        type: string
      expires_in_seconds:
        type: integer
      password:
        type: string
    type: object
  api.LoginResponse:
    properties:
      email:
        type: string
      id:
        type: string
      refresh_token:
        type: string
      token:
        type: string
    type: object
  api.RefreshResponse:
    properties:
      token:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.LoginParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: login user
      tags:
      - auth
...
  /refresh:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: revoke refresh token
      tags:
      - auth
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.RefreshResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: refresh access token
      tags:
      - auth
...
swagger: "2.0"